+++ src/snyk/common/proxy.ts
@@ -1,5 +1,5 @@
 import { AxiosRequestConfig } from 'axios';
-import { HttpsProxyAgent } from 'https-proxy-agent';
+import createHttpsProxyAgent, { HttpsProxyAgent, HttpsProxyAgentOptions } from 'https-proxy-agent';
 import * as url from 'url';
 import { IVSCodeWorkspace } from './vscode/workspace';
 
@@ -7,6 +7,17 @@ export function getHttpsProxyAgent(
   workspace: IVSCodeWorkspace,
   processEnv: NodeJS.ProcessEnv = process.env,
 ): HttpsProxyAgent | undefined {
+  const proxyOptions = getProxyOptions(workspace, processEnv);
+  if (proxyOptions == undefined) {
+    return undefined;
+  }
+  return new HttpsProxyAgent(proxyOptions);
+}
+
+export function getProxyOptions(
+  workspace: IVSCodeWorkspace,
+  processEnv: NodeJS.ProcessEnv = process.env,
+): HttpsProxyAgentOptions | undefined {
   let proxy: string | undefined = getVsCodeProxy(workspace);
   if (!proxy) {
     proxy = processEnv.HTTPS_PROXY || processEnv.https_proxy || processEnv.HTTP_PROXY || processEnv.http_proxy;
@@ -21,15 +32,18 @@ export function getHttpsProxyAgent(
     return undefined;
   }
 
+  if (proxyUrl.hostname == null || proxyUrl.hostname === '' || proxyUrl.port == null || proxyUrl.port === '') {
+    return undefined;
+  }
+
   const strictProxy = workspace.getConfiguration<boolean>('http', 'proxyStrictSSL') ?? true;
-  const proxyOptions = {
+
+  return {
     host: proxyUrl.hostname,
-    port: proxyUrl.port ? parseInt(proxyUrl.port, 10) : null,
+    port: parseInt(proxyUrl.port, 10),
     auth: proxyUrl.auth,
     rejectUnauthorized: strictProxy,
   };
-
-  return new HttpsProxyAgent(proxyOptions);
 }
 
 export function getVsCodeProxy(workspace: IVSCodeWorkspace): string | undefined {
@@ -42,4 +56,14 @@ export function getAxiosProxyConfig(workspace: IVSCodeWorkspace): AxiosRequestCo
     httpAgent: getHttpsProxyAgent(workspace),
     httpsAgent: getHttpsProxyAgent(workspace),
   };
+}
+
+export function getProxyEnvVariable(proxyOptions: createHttpsProxyAgent.HttpsProxyAgentOptions | undefined): string {
+  if (!proxyOptions) {
+    return '';
+  }
+  const { host, port } = proxyOptions;
+  // TODO: add auth to proxy env variable
+  // noinspection HttpUrlsUsage
+  return `http://${host}:${port}`;
 }