+++ src/test/unit/common/configuration.test.ts
@@ -9,6 +9,7 @@ import SecretStorageAdapter from '../../../snyk/common/vscode/secretStorage';
 import { ExtensionContext } from '../../../snyk/common/vscode/types';
 import { IVSCodeWorkspace } from '../../../snyk/common/vscode/workspace';
 import { stubWorkspaceConfiguration } from '../mocks/workspace.mock';
+import { extensionContextMock } from '../mocks/extensionContext.mock';
 
 suite('Configuration', () => {
   let workspaceStub: IVSCodeWorkspace;
@@ -19,13 +20,7 @@ suite('Configuration', () => {
 
     let token = '';
 
-    extensionContext = {
-      secrets: {
-        store: (_key: string, _value: string) => Promise.resolve(),
-        get: () => Promise.resolve(),
-        delete: () => Promise.resolve(),
-      },
-    } as unknown as ExtensionContext;
+    extensionContext = extensionContextMock;
     SecretStorageAdapter.init(extensionContext);
 
     const stub = sinon.stub().returns({